PDI-10310 : Fix all -Xlint:all warnings in PDI
git-svn-id: svn://source.pentaho.org/svnkettleroot/Kettle/trunk@18795 5fb7f6ec-07c1-534a-b4ca-9155e429e800

+++ engine/src/org/pentaho/di/trans/steps/denormaliser/Denormaliser.java
@@ -238,8 +238,8 @@ public class Denormaliser extends BaseStep implements StepInterface
                 Object sum  = data.sum[i];
                 if (count>0)
                 {
-                	if (sum instanceof Long) resultValue = (long)((Long)sum / count);
-                	else if (sum instanceof Double) resultValue = (double)((Double)sum / count);
+                	if (sum instanceof Long) resultValue = (Long)sum / count;
+                	else if (sum instanceof Double) resultValue = (Double)sum / count;
                 	else if (sum instanceof BigDecimal) resultValue = ((BigDecimal)sum).divide(new BigDecimal(count));
                 	else resultValue = null; // TODO: perhaps throw an exception here?<
                 }
